# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stockcast API",
    "version": "0.7.0"
  },
  "paths": {
...
    "/forecast-runs": {
      "post": {
        "tags": [
          "forecasts"
        ],
        "summary": "Trigger Forecast",
        "operationId": "trigger_forecast_forecast_runs_post",
        "parameters": [
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 90,
              "title": "Horizon"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Org-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "forecasts"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_forecast_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Org-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ForecastRunRead"
                  },
                  "title": "Response List Runs Forecast Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/forecasts": {
      "get": {
        "tags": [
          "forecasts"
        ],
        "summary": "Get Forecast",
        "operationId": "get_forecast_forecasts_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 90,
              "title": "Days"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Org-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductForecast"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/forecast-accuracy": {
      "get": {
        "tags": [
          "forecasts"
        ],
        "summary": "Accuracy",
        "operationId": "accuracy_forecast_accuracy_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "worst",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 10,
              "title": "Worst"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Org-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccuracySummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/forecasts/simulate": {
      "post": {
        "tags": [
          "forecasts"
        ],
        "summary": "Simulate Promotion",
        "description": "What-if: apply a draft promotion to the latest forecast; returns unit delta per product\n(incl. the post-promo dip) and the raw-material impact via BOM explosion.",
        "operationId": "simulate_promotion_forecasts_simulate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Org-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccuracyRow": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "history_days": {
            "type": "integer",
            "title": "History Days"
          },
          "zero_share": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Zero Share"
          },
          "holdout_days": {
            "type": "integer",
            "title": "Holdout Days"
          },
          "mape": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mape"
          },
          "wape": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wape"
          },
          "wape_chronos": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wape Chronos"
          },
          "wape_stats": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wape Stats"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "model",
          "history_days",
          "zero_share",
          "holdout_days",
          "mape",
          "wape",
          "wape_chronos",
          "wape_stats"
        ],
        "title": "AccuracyRow"
      },
      "AccuracySummary": {
        "properties": {
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "skus": {
            "type": "integer",
            "title": "Skus"
          },
          "wape": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wape"
          },
          "median_sku_wape": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Sku Wape"
          },
          "by_model": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Model"
          },
          "worst": {
            "items": {
              "$ref": "#/components/schemas/AccuracyRow"
            },
            "type": "array",
            "title": "Worst"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "as_of",
          "skus",
          "wape",
          "median_sku_wape",
          "by_model",
          "worst"
        ],
        "title": "AccuracySummary"
      },
...
      "ChannelShare": {
        "properties": {
          "channel_id": {
            "type": "string",
            "format": "uuid",
            "title": "Channel Id"
          },
          "channel_name": {
            "type": "string",
            "title": "Channel Name"
          },
          "channel_type": {
            "type": "string",
            "title": "Channel Type"
          },
          "share": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Share"
          },
          "units_90d": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Units 90D"
          },
          "p50_30d": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "P50 30D"
          },
          "p50_90d": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "P50 90D"
          }
        },
        "type": "object",
        "required": [
          "channel_id",
          "channel_name",
          "channel_type",
          "share",
          "units_90d",
          "p50_30d",
          "p50_90d"
        ],
        "title": "ChannelShare"
      },
...
      "ForecastPoint": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "p10": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "P10"
          },
          "p50": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "P50"
          },
          "p90": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "P90"
          },
          "factor": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Factor"
          },
          "event": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event"
          }
        },
        "type": "object",
        "required": [
          "date",
          "p10",
          "p50",
          "p90",
          "factor",
          "event"
        ],
        "title": "ForecastPoint"
      },
      "ForecastRunRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "skus_total": {
            "type": "integer",
            "title": "Skus Total"
          },
          "skus_chronos": {
            "type": "integer",
            "title": "Skus Chronos"
          },
          "skus_croston": {
            "type": "integer",
            "title": "Skus Croston"
          },
          "skus_fallback": {
            "type": "integer",
            "title": "Skus Fallback"
          },
          "wape": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wape"
          },
          "wape_base": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wape Base"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "org_id",
          "status",
          "trigger",
          "horizon_days",
          "as_of",
          "started_at",
          "finished_at",
          "skus_total",
          "skus_chronos",
          "skus_croston",
          "skus_fallback",
          "wape",
          "wape_base",
          "error"
        ],
        "title": "ForecastRunRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialDeltaRead": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "delta_qty": {
            "type": "number",
            "title": "Delta Qty"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "sku",
          "unit",
          "delta_qty"
        ],
        "title": "MaterialDeltaRead"
      },
...
      "ProductDeltaRead": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "base_units": {
            "type": "number",
            "title": "Base Units"
          },
          "promo_units": {
            "type": "number",
            "title": "Promo Units"
          },
          "delta_units": {
            "type": "number",
            "title": "Delta Units"
          },
          "post_dip_units": {
            "type": "number",
            "title": "Post Dip Units"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "sku",
          "base_units",
          "promo_units",
          "delta_units",
          "post_dip_units"
        ],
        "title": "ProductDeltaRead"
      },
      "ProductForecast": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ForecastPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "total_p50_30d": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total P50 30D"
          },
          "total_p50_90d": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total P50 90D"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelShare"
            },
            "type": "array",
            "title": "Channels",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "run_id",
          "as_of",
          "model",
          "points",
          "total_p50_30d",
          "total_p50_90d"
        ],
        "title": "ProductForecast"
      },
...
      "PromotionScope": {
        "type": "string",
        "enum": [
          "all",
          "category",
          "skus"
        ],
        "title": "PromotionScope"
      },
      "PromotionType": {
        "type": "string",
        "enum": [
          "paid_ads",
          "discount",
          "email"
        ],
        "title": "PromotionType"
      },
...
      "SimulateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/PromotionType"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "discount_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Pct"
          },
          "spend_amount": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spend Amount"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "scope": {
            "$ref": "#/components/schemas/PromotionScope",
            "default": "all"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "product_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Ids"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "start_date",
          "end_date"
        ],
        "title": "SimulateRequest"
      },
      "SimulateResponse": {
        "properties": {
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "lift": {
            "type": "number",
            "title": "Lift"
          },
          "post_dip": {
            "type": "number",
            "title": "Post Dip"
          },
          "total_delta_units": {
            "type": "number",
            "title": "Total Delta Units"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductDeltaRead"
            },
            "type": "array",
            "title": "Products"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialDeltaRead"
            },
            "type": "array",
            "title": "Materials"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "lift",
          "post_dip",
          "total_delta_units",
          "products",
          "materials"
        ],
        "title": "SimulateResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}